simple_lmk: Introduce Simple Low Memory Killer for Android
This is a complete low memory killer solution for Android that is small
and simple. Processes are killed according to the priorities that
Android gives them, so that the least important processes are always
killed first. Processes are killed until memory deficits are satisfied,
as observed from the page allocator's slow allocation path.

The only tunable is the desired amount of memory to be freed per reclaim
event. Simple LMK tries to free at least the desired amount of memory
and waits until all of its victims' memory is freed before proceeding to
kill more processes.

Simple LMK also comes with support for the OOM reaper feature, if the
kernel in use supports it. It can be enabled using the Kconfig option.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>